requirements.txt
loguru==0.7.2
pyinstaller==6.9.0
python-dotenv==1.0.1
bilibili-api-python==16.3.0